feat: 新增POSIX内存映射文件实现并移除PE工具Windows依赖

+++ src/resolver.cpp
@@ -358,7 +358,7 @@ void Resolver::ProcessModules(
                                         code_size,
                                         section_offset,
                                         current.linenumStart,
-                                        current.linenumStart + current.deltaLineEnd,
+                                        static_cast<std::uint32_t>(current.linenumStart + current.deltaLineEnd),
                                         section_index,
                                         nullptr,
                                     });
